force type hint error

+++ very_basic_calculator/power.py
@@ -1,2 +1,2 @@
-def power(x: float, y: int) -> float:
-    return int(x ** y)
+def power(x: float, y: int) -> int:
+    return x ** y
